img-gallery/requirements.txt
Pillow
tqdm
fastapi
uvicorn
Jinja2
imagehash